attempts to make eslint happy, not entirely successfully
The original problem was that the eslint playwright plugin
was finding issues with the RTL tests. So I added an overrides
to the eslint config file to address that.
That worked to a degree, as eslint is no longer reporting
issues with the RTL tests. It does result in eslint now having
an issue with useInputMask.ts on line 19. It's unclear to me
why this did not happen with the old configuration, but does
happen with the new one, i.e. after adding the overrides.