exclude msw from vite dep optimization
Playwright tests were failing, depending on which ones you run
and on the number of Playwright workers. A consistent way to
reproduce the problem was to run all tests with 3 Playwright workers.
This caused the app e2e tests to fail beacuse of a 504 Outdated
Optimize Dep response from node_modules/.vite/deps/msw.js.

Excluding msw from the vite dep optimization seems to fix this issue.
Note that using the --force flag or config option for vite dep
optimization, might improve the situation, but didn't solve it.