Update GitHub Actions workflows triggers
Do not trigger Frontend and Backend workflows on push, but only for pull
request and merge queue (merge group). Because commits to `main` are
already checked in the merge group, they do not need to run on push.
Also, the `pull_request` trigger will run on the PR merge branch instead
of the PR branch, which lets us find potential test failures sooner
(in the PR of only when the merge queue checks run).