Restructure election and polling station fixtures
Move into a fixture per election, with both the election itself and its
polling stations. This way, not all elections and polling stations have
to be loaded for all tests.

`election_2.sql` can be skipped in the default seeded fixtures because
it is not used by Playwright d2d tests.