Corrected to state highest instead of largest
A bit of nitpicking :grin: But because largest is used in combination with remainder and highest in combination with averages it seems fair to do change the wording here.

Interestingly the condition/method call made here this comment applies to seems to be not supported by law. In P8 3 it only mentions that only 1 seat should be assigned. No mention what to do in case all parties have 1 seat and there would still be seats left. While here in the code it just implements to use the 'normal' way to assign. So unless there is jurisdictional evidence this has been done in the past, I think the correct way to implement here is to throw an error and that the kiesraad should decide what to do. (I also think, in practice, this  case will never occur, so it's a bit of a theoretical case. But it should be handled supported by law).